finish

+++ app/Http/Controllers/TasksController.php
@@ -48,10 +48,15 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             ]);
         
-        $task = new Task;
+        /*$task = new Task;
         $task->content = $request->content;
         $task->status = $request->status;
-        $task->save();
+        $task->save();*/
+        
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
         
         return redirect('/');
         //
@@ -101,9 +106,11 @@ class TasksController extends Controller
             ]);
         
         $task = Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id){
         $task->content = $request->content;
         $task->status = $request->status;
         $task->save();
+        }
         
         return redirect('/');
         //
@@ -118,8 +125,9 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id){
         $task->delete();
-        
+        }
         return redirect('/');
         //
     }
